updates to waybar styles

+++ .config/hypr/scripts/start-waybar.sh
@@ -1,4 +1,8 @@
 #! /bin/bash
 
-killall waybar
-waybar -c $HOME/.config/hypr/waybar/config.jsonc -s $HOME/.config/hypr/waybar/style.css 
+#kill waybar if running
+if [ ! -z "$(pidof waybar)" ]; then
+    killall waybar
+fi
+
+waybar -c $HOME/.config/hypr/waybar/config.jsonc -s $HOME/.config/hypr/waybar/style.scss & disown
